+++ tests/admin.rs
@@ -1,8 +1,8 @@
+use holochain::test_utils::itertools::Itertools;
 use holochain::{prelude::AppBundleSource, sweettest::SweetConductor};
 use holochain_client::{AdminWebsocket, AppWebsocket, InstallAppPayload, InstalledAppId};
-use holochain_conductor_api::CellInfo;
+use holochain_conductor_api::{CellInfo, StorageBlob};
 use holochain_zome_types::ExternIO;
-use rand::Rng;
 use std::{collections::HashMap, path::PathBuf};
 use utilities::{authorize_signing_credentials, sign_zome_call};
 
@@ -26,13 +26,8 @@ async fn signed_zome_call() {
     let mut admin_ws = AdminWebsocket::connect(format!("ws://localhost:{}", admin_port))
         .await
         .unwrap();
-
-    let mut rng = rand::thread_rng();
-    let random_number: u8 = rng.gen();
-    let app_id: InstalledAppId = format!("test-app-{}", random_number).into();
-
+    let app_id: InstalledAppId = "test-app".into();
     let agent_key = admin_ws.generate_agent_pub_key().await.unwrap();
-
     admin_ws
         .install_app(InstallAppPayload {
             agent_key: agent_key.clone(),
@@ -43,14 +38,11 @@ async fn signed_zome_call() {
         })
         .await
         .unwrap();
-
     admin_ws.enable_app(app_id.clone()).await.unwrap();
-
     let app_ws_port = admin_ws.attach_app_interface(30000).await.unwrap();
     let mut app_ws = AppWebsocket::connect(format!("ws://localhost:{}", app_ws_port))
         .await
         .unwrap();
-
     let installed_app = app_ws.app_info(app_id).await.unwrap().unwrap();
 
     let cells = installed_app.cell_info.into_values().next().unwrap();
@@ -78,4 +70,64 @@ async fn signed_zome_call() {
         ExternIO::decode::<String>(&response).unwrap(),
         TEST_FN_NAME.to_string()
     );
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn storage_info() {
+    let conductor = SweetConductor::from_standard_config().await;
+    let admin_port = conductor.get_arbitrary_admin_websocket_port().unwrap();
+    let mut admin_ws = AdminWebsocket::connect(format!("ws://localhost:{}", admin_port))
+        .await
+        .unwrap();
+    let app_id: InstalledAppId = "test-app".into();
+    let agent_key = admin_ws.generate_agent_pub_key().await.unwrap();
+    admin_ws
+        .install_app(InstallAppPayload {
+            agent_key: agent_key.clone(),
+            installed_app_id: Some(app_id.clone()),
+            membrane_proofs: HashMap::new(),
+            network_seed: None,
+            source: AppBundleSource::Path(PathBuf::from("./fixture/test.happ")),
+        })
+        .await
+        .unwrap();
+    admin_ws.enable_app(app_id.clone()).await.unwrap();
+
+    let storage_info = admin_ws.storage_info().await.unwrap();
+
+    let matched_storage_info = storage_info
+        .blobs
+        .iter()
+        .filter(|b| match b {
+            StorageBlob::Dna(dna_storage_info) => dna_storage_info.used_by.contains(&app_id),
+            _ => false,
+        })
+        .collect_vec();
+    assert_eq!(1, matched_storage_info.len());
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn dump_network_stats() {
+    let conductor = SweetConductor::from_standard_config().await;
+    let admin_port = conductor.get_arbitrary_admin_websocket_port().unwrap();
+    let mut admin_ws = AdminWebsocket::connect(format!("ws://localhost:{}", admin_port))
+        .await
+        .unwrap();
+    let app_id: InstalledAppId = "test-app-".into();
+    let agent_key = admin_ws.generate_agent_pub_key().await.unwrap();
+    admin_ws
+        .install_app(InstallAppPayload {
+            agent_key: agent_key.clone(),
+            installed_app_id: Some(app_id.clone()),
+            membrane_proofs: HashMap::new(),
+            network_seed: None,
+            source: AppBundleSource::Path(PathBuf::from("./fixture/test.happ")),
+        })
+        .await
+        .unwrap();
+    admin_ws.enable_app(app_id.clone()).await.unwrap();
+
+    let network_stats = admin_ws.dump_network_stats().await.unwrap();
+
+    assert!(network_stats.contains("\"backend\": \"go-pion\""));
 }