+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+- Add `storage_info` to the admin websocket.
+- Add `network_info` to the app websocket.
 ### Changed
+- Upgrade to Holochain 0.2 release candidate ahead of the holochain 0.2 release.
 ### Fixed
 ### Removed